# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Collab API
  version: 0.1.0
  description: Local-first collaboration platform with RAG knowledge QA and AI task planning.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /admin/models:
    get:
      tags: [Admin]
      summary: List model configurations
      responses:
        '200':
          description: API keys are never returned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiMeta'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/ModelConfigurationView' }
    post:
      tags: [Admin]
      summary: Create a model configuration
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ModelConfigurationRequest' }
      responses:
        '201':
          description: Model configuration created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiMeta'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/ModelConfigurationView' }
  /admin/models/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      tags: [Admin]
      summary: Update a model configuration
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ModelConfigurationRequest' }
      responses:
        '200':
          description: Model configuration updated
    delete:
      tags: [Admin]
      summary: Delete a model configuration
      responses:
        '204': { description: Deleted }
  /admin/models/{id}/test:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Admin]
      summary: Test a model connection
      responses:
        '200': { description: Provider response and normalized usage returned }
...
  /admin/models/assignments/{purpose}:
    parameters:
      - name: purpose
        in: path
        required: true
        schema: { $ref: '#/components/schemas/ModelPurpose' }
    put:
      tags: [Admin]
      summary: Assign an enabled configuration to a purpose
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [configurationId]
              properties:
                configurationId: { type: string, format: uuid }
      responses:
        '204': { description: Assignment updated }
...
  /projects:
    get:
      tags: [Projects]
      summary: List current user's projects
      description: Uses one membership-scoped join; it does not issue one role query per project.
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectEnvelope'
        '400':
          description: Project request validation failed
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /projects/{projectId}/documents:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Documents]
      summary: List documents
      description: Project members can list documents, ordered by createdAt descending. Object keys, chunk content, and vectors are never returned.
      responses:
        '200':
          description: Document list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListEnvelope'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Documents]
      summary: Upload document
      description: >
        OWNER and ADMIN can upload one validated PDF, DOCX, Markdown, or TXT file.
        Validation and the private MinIO put happen outside the database transaction.
        PREPARING and ACTIVE projects are writable; COMPLETED and ARCHIVED projects
        return PROJECT_READ_ONLY. A short registration transaction locks the project row,
        checks its status and the 100-document limit, inserts metadata and audit, and
        publishes the after-commit event. If registration fails, the already-written object
        is compensating-deleted.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                displayName:
                  type: string
                  maxLength: 180
      responses:
        '202':
          description: Accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Project is read-only (PROJECT_READ_ONLY), or the document limit was reached
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /projects/{projectId}/documents/{documentId}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/DocumentId'
    get:
      tags: [Documents]
      summary: Get document
      description: A project member can read safe metadata. Cross-project and non-member access use the same 404 response.
      responses:
        '200':
          description: Document detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentEnvelope'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Documents]
      summary: Delete document and its chunks
      description: OWNER and ADMIN only. PREPARING and ACTIVE projects are writable; COMPLETED and ARCHIVED return PROJECT_READ_ONLY. Deletes the private object, chunks, citations through foreign keys, and document row.
      responses:
        '204':
          description: Deleted
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /projects/{projectId}/documents/{documentId}/download-url:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/DocumentId'
    get:
      tags: [Documents]
      summary: Create five-minute presigned download URL
      description: Project members receive an ephemeral URL for the private bucket. DELETING documents cannot be downloaded.
      responses:
        '200':
          description: Download URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadUrlEnvelope'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /projects/{projectId}/documents/{documentId}/retry:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/DocumentId'
    post:
      tags: [Documents]
      summary: Retry failed processing
      description: >
        OWNER and ADMIN only. PREPARING and ACTIVE projects are writable; COMPLETED and
        ARCHIVED return PROJECT_READ_ONLY. Resets the same FAILED document to UPLOADED and reuses its
        existing private object. A new processing token is generated when the worker claims
        the retry; an older worker cannot heartbeat, fail, replace chunks, or complete it.
      responses:
        '202':
          description: Retry accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentEnvelope'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /projects/{projectId}/documents/{documentId}/reindex:
    post:
      tags: [Documents]
      summary: Reindex one ready document
      description: PREPARING and ACTIVE projects are writable; COMPLETED and ARCHIVED return PROJECT_READ_ONLY.
      responses:
        '202': { description: Reindex accepted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /projects/{projectId}/documents/reindex-all:
    post:
      tags: [Documents]
      summary: Reindex all ready documents as project administrator
      description: PREPARING and ACTIVE projects are writable; COMPLETED and ARCHIVED return PROJECT_READ_ONLY.
      responses:
        '202': { description: Batch accepted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /projects/{projectId}/documents/reindex-progress:
    get:
      tags: [Documents]
      summary: Get in-memory batch reindex queue progress
      responses:
        '200': { description: Batch progress }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    DocumentId:
      name: documentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Invalid request or file content
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Unauthorized:
      description: Unauthorized
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Forbidden:
      description: Forbidden
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    NotFound:
      description: Not found
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Conflict:
      description: Conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    PayloadTooLarge:
      description: Payload too large
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    UnsupportedMediaType:
      description: Unsupported media type
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    ServiceUnavailable:
      description: Private object storage or configured Chat service is unavailable
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
  schemas:
    ApiMeta:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiMeta'
        - type: object
          required: [data]
          properties:
            data: { type: 'null' }
...
    ProjectRole:
      type: string
      enum: [OWNER, ADMIN, MEMBER]
    ProjectType:
      type: string
      enum: [COMPETITION, COURSE_DESIGN, SOFTWARE_TRAINING, OTHER]
    ProjectView:
      type: object
      required: [id, name, type, status, role, ownerId, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        ownerId: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/ProjectType' }
        startDate: { type: [string, 'null'], format: date }
        dueDate: { type: [string, 'null'], format: date }
        status: { $ref: '#/components/schemas/ProjectStatus' }
        role: { $ref: '#/components/schemas/ProjectRole' }
        version: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProjectStatus:
      type: string
      enum: [PREPARING, ACTIVE, COMPLETED, ARCHIVED]
    CreateProjectRequest:
      type: object
      required: [name, type]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, maxLength: 2000, default: '' }
        type: { $ref: '#/components/schemas/ProjectType' }
        startDate: { type: [string, 'null'], format: date }
        dueDate: { type: [string, 'null'], format: date }
...
    ProjectEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiMeta'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ProjectView' }
    ProjectListEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiMeta'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/ProjectView' }
...
    ModelProviderType:
      type: string
      enum: [OPENAI_COMPATIBLE, ANTHROPIC, GEMINI]
    ModelPurpose:
      type: string
      enum: [KNOWLEDGE_CHAT, AGENT, PLANNING]
    ModelCapability:
      type: string
      enum: [CHAT, STREAMING, STRUCTURED_OUTPUT, NATIVE_TOOLS, USAGE]
    ModelConfigurationRequest:
      type: object
      required:
        [name, providerType, baseUrl, apiPath, modelName, enabled,
         temperature, maxOutputTokens, capabilities]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        providerType: { $ref: '#/components/schemas/ModelProviderType' }
        baseUrl: { type: string, format: uri, maxLength: 2048 }
        apiPath: { type: string, minLength: 1, maxLength: 512 }
        apiKey: { type: [string, 'null'], writeOnly: true, maxLength: 1000 }
        modelName: { type: string, minLength: 1, maxLength: 160 }
        enabled: { type: boolean }
        temperature: { type: number, minimum: 0, maximum: 2 }
        maxOutputTokens: { type: integer, minimum: 1, maximum: 131072 }
        capabilities:
          type: array
          minItems: 1
          uniqueItems: true
          items: { $ref: '#/components/schemas/ModelCapability' }
    ModelConfigurationView:
      type: object
      required:
        [id, name, providerType, baseUrl, apiPath, hasApiKey, modelName,
         enabled, temperature, maxOutputTokens, capabilities, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        providerType: { $ref: '#/components/schemas/ModelProviderType' }
        baseUrl: { type: string, format: uri }
        apiPath: { type: string }
        hasApiKey: { type: boolean }
        modelName: { type: string }
        enabled: { type: boolean }
        temperature: { type: number }
        maxOutputTokens: { type: integer }
        capabilities:
          type: array
          items: { $ref: '#/components/schemas/ModelCapability' }
        updatedAt: { type: string, format: date-time }
...
    DocumentStatus:
      type: string
      enum: [UPLOADED, PARSING, INDEXING, READY, FAILED, DELETING]
    DocumentView:
      type: object
      description: Safe document metadata. objectKey, extracted text, chunks, vectors, credentials, and server paths are intentionally absent.
      required: [id, projectId, displayName, originalFilename, mimeType, sizeBytes, status, chunkCount, uploadedById, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        displayName: { type: string }
        originalFilename: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: integer, format: int64 }
        status: { $ref: '#/components/schemas/DocumentStatus' }
        parserType: { type: [string, 'null'] }
        chunkCount: { type: integer }
        embeddingProvider: { type: [string, 'null'] }
        embeddingModel: { type: [string, 'null'] }
        embeddingDimension: { type: [integer, 'null'] }
        errorMessage: { type: [string, 'null'] }
        uploadedById: { type: string, format: uuid }
        uploadedByDisplayName: { type: [string, 'null'] }
        indexedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DocumentEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiMeta'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/DocumentView' }
    DocumentListEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiMeta'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/DocumentView' }
    DownloadUrlEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiMeta'
        - type: object
          properties:
            data:
              type: object
              required: [url, expiresAt]
              properties:
                url: { type: string, format: uri }
                expiresAt: { type: string, format: date-time }
...